requirements.txt
dash~=2.6.2
pandas~=1.5.1
PyYAML~=6.0
dash-bootstrap-components~=1.2.1
dash-bootstrap-templates~=1.0.7
